Cleaned data
Added dummy variables for each season, changed holiday and functioning day from qualitative data to 1s and 0s.

For each season a 1 means that observation was during that season, a 0 means it was not.

For "Holiday" a 1 means that it was a holiday, a 0 means it was not.

For "Functioning Day" a 1 means it was a functioning day, a 0 means it was not.